Add files via upload
The generalised Kelly Criterion (a.k.a. The Real Kelly) for independent, concurrent events.

An implementation of the generalised Kelly Criterion (a.k.a. The Real Kelly) for independent, concurrent events as discussed in this article https://www.pinnacle.com/en/betting-articles/Betting-Strategy/the-real-kelly-criterion/HZKJTFCB3KNYN9CJ

The algorithm will find optimal bet sizes for a set of concurrent singles and/or 'round robin' combinations of parlays or teasers.